rename storage schemas exposed via webserver
